:truck: Reorganize CSS files and style rules
Rename the CSS `templates` directory to `layout` to be more
specific to what it contains.

Rename "global" to "environment" and move component-specific
styles into the "components" sheet so that the environment rules
only include broadly applicable, high-level rules. Also move icon
styles to the new environment sheet.

Move all of the CSS rules in the `elements` stylesheet part into
either the `environment` or the `components` parts, as applicable.
Then removed the emptied elements stylesheet and directory. This
will help prevent confusion from having image-related rules, for
example, in multiple places.