:art: Switch to max-width on main content
Switch from a blanket 70% width to Switch to 100% width with a max
width to be more flexible and make better use of space on
medium-width viewports. Before a lot of space would be wasted from
forcing 70% on viewport widths that could accessibly accommodate
more width.

This required adding a div wrapper to the article content body on
single pages (post pages already had it) but allowed us to merge
the single page and post styles.

Also updated `alignwide` and `alignfull` styles to match, using
the (not perfect) 100vw approach to assure "wide" and "full" width
beyond the content max.