Fix #100 Use local path to load styles more reliably
The problem seemed to stem from a malformed path to the editor
stylesheet. The URI was built using `get_stylesheet_directory_uri`,
which worked everywhere but one environment. Switching from that
to a plain relative URL in `add_editor_style` seemed to clear up
the problem and load editor styles in all environments.